Split sql file into individual model files
